only the user who added a brand can edit or delete it

+++ controllers/brands.js
@@ -29,7 +29,7 @@ router.get('/new', isSignedin, (req, res) => {
 //Create route
 router.post('/', isSignedin, async (req, res) => {
     try {
-        
+
         await Brand.create(req.body)
         return res.redirect('/brands')
     } catch (error) {
@@ -42,9 +42,8 @@ router.post('/', isSignedin, async (req, res) => {
 router.get('/:brandId', async (req, res, next) => {
     try {
         if (mongoose.Types.ObjectId.isValid(req.params.brandId)) {
-            const brand = await Brand.findById(req.params.brandId)
+            const brand = await Brand.findById(req.params.brandId).populate('addedBy')
             if (!brand) return next()
-            console.log(brand)
             return res.render('brands/show.ejs', { brand })
         } else {
             next()
@@ -56,11 +55,19 @@ router.get('/:brandId', async (req, res, next) => {
 })
 
 //Edit page - form
-router.get('/:brandId/edit', async (req, res) => {
+router.get('/:brandId/edit', isSignedin, async (req, res, next) => {
     try {
-        const brand = await Brand.findById(req.params.brandId)
-        console.log(brand)
-        return res.render('brands/edit.ejs', { brand })
+        if (mongoose.Types.ObjectId.isValid(req.params.brandId)) {
+            const brand = await Brand.findById(req.params.brandId)
+            if (!brand) return next()
+            if (!brand.addedBy.equals(req.session.user._id)) {
+                return res.redirect(`/brands/${req.params.brandId}`)
+            }
+            return res.render('brands/edit.ejs', { brand })
+        } else {
+            next()
+        }
+
     } catch (error) {
         console.log(error)
         return res.status(500).send('An error occurred')
@@ -68,10 +75,14 @@ router.get('/:brandId/edit', async (req, res) => {
 })
 
 //Update route
-router.put('/:brandId', async (req, res) => {
+router.put('/:brandId', isSignedin, async (req, res) => {
     try {
-        await Brand.findByIdAndUpdate(req.params.brandId, req.body)
-        return res.redirect(`/brands/${req.params.brandId}`)
+        const brandToUpdate = await Brand.findById(req.params.brandId)
+        if (brandToUpdate.addedBy.equals(req.session.user._id)) {
+            await Brand.findByIdAndUpdate(req.params.brandId, req.body)
+            return res.redirect(`/brands/${req.params.brandId}`)
+        }
+        throw new Error('You are not authorised to perform this action')
     } catch (error) {
         console.log(error)
         return res.status(500).send('An error occurred')
@@ -81,9 +92,12 @@ router.put('/:brandId', async (req, res) => {
 //Delete route
 router.delete('/:brandId', async (req, res) => {
     try {
-        const brandToDelete = await Brand.findByIdAndDelete(req.params.brandId)
-        console.log(brandToDelete)
-        return res.redirect('/brands')
+        const brandToDelete = await Brand.findById(req.params.brandId)
+        if (brandToDelete.addedBy.equals(req.session.user._id)) {
+            await Brand.findByIdAndDelete(req.params.brandId)
+            return res.redirect('/brands')
+        }
+        throw new Error('You are not authorised to perform this action')
     } catch (error) {
         console.log(error)
         return res.status(500).send('An error occurred')
